freshservice_scraper/requirements.txt
requests==2.31.0
freshpy
python-dateutil==2.8.2
pandas==2.0.0
pendulum==2.1.2
pytest
pytest-cov
